+++ cubi_sak/exceptions.py
@@ -7,3 +7,7 @@ class CubiSakWarning(Warning):
 
 class CubiSakException(Exception):
     """Base ``Exception`` class."""
+
+
+class UnsupportedIsaTabFeaturException(CubiSakException):
+    """Raised when an unsupported ISA-tab feature occurs."""