+++ HISTORY.md
@@ -5,6 +5,7 @@
 - Adjusting package meta data in `setup.py`.
 - Fixing documentation bulding bug.
 - Documentation is now built during testing.
+- Adding `cubi-sak snappy pull-sheet`.
 
 ## v0.1.0